Refs #322: Further improvements to make file saves work more intuitively.
* Added three NeXus object flags, 'infile', 'saved', and 'changed' to record whether
the NXobject currently exists in the file, whether its current value has been saved,
and whether it has been changed. The third flag is for third-party scripts that need
to keep track of when a variable has been changed. It can be reset using the 
'set_unchanged' method.

* Added write methods to the NXobjects to write the object directly to the NeXus file. 
This required moving the context routines (__enter__ and __exit__) to the NXobject 
class. The use of the 'with ... as ...' syntax means that the API now requires 
Python 2.5.

* Used the new 'walk' methods in conjunction with the 'write' methods to allow
incremental saves (when no file name is specified) without needing to reload the
entire file.

* Debugged the reading and writing of NXlink objects to NeXus files. 



git-svn-id: https://svn.nexusformat.org/code/trunk@1806 ff5d1e40-2be0-497f-93bd-dc18237bd3c7