Merge in changes from trunk r1924
Fix a couple of issues involving a NeXus HDF5 when no class is specified,
in particular XML must have a class name as it is used as the XML node
name. For XML will use NXcollection if no class is specified. Also
NXgetgroupinfo() did not work correctly in the case of no class, causing 
NXtraverse to fail.

Also fix XML issue when an attribute contains text that looks like
a nexus data type - the code previously complained about a missing ":"

The problem with NXdir giving an attribute error when parsing an externally
linked file is also fixed (in NXgetnextattr)

refs #306



git-svn-id: https://svn.nexusformat.org/code/branches/4.3@1915 ff5d1e40-2be0-497f-93bd-dc18237bd3c7