First Commit of c-nxvalidate
This is the first version which recurses through a file. Many tests and feature  are not implemented.
But from this version on, other people can sensibly help out. Just search for TODO and talk to Mark K.
I am also still looking for an elegant way to validate attributes. But may be, there is node.
Refers to #363